update return type

+++ src/dice.rs
@@ -1,6 +1,6 @@
 use rand::prelude::*;
 
-pub fn roll_dice() -> u32 {
+pub fn roll_dice() -> usize {
     let mut rng = rand::thread_rng();
     rng.gen_range(1..=6)
 }
